backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-binance==1.0.19
pandas-ta
pandas
numpy
python-dotenv==1.0.0
websockets==12.0
asyncio==3.4.3
setuptools
wheel
